Changed file structure of dashboard components, and extracted render logic into separate files from DashboardLogDisplay
